Unify model classes so that all business logic related to a class resides in one file.
