+++ app/shared/models/project.rb
@@ -1,6 +1,7 @@
 require "securerandom"
 
 require_relative "../../shared/logging_module"
+require_relative "../json_convertible"
 require_relative "artifact_provider"
 require_relative "local_artifact_provider"
 require_relative "job_trigger"
@@ -9,6 +10,7 @@ module FastlaneCI
   # All metadata about a project.
   # A project is usually a git url, a lane, and a project name, like "Production Build"
   class Project
+    include FastlaneCI::JSONConvertible
     include FastlaneCI::Logging
 
     # @return [RepoConfig] Repo configuration
@@ -71,6 +73,49 @@ module FastlaneCI
 
     def local_repo_path
       return File.join(File.expand_path("~/.fastlane/ci/"), id)
+    end
+
+    def self.attribute_to_type_map
+      return { :@repo_config => GitHubRepoConfig }
+    end
+
+    def self.map_enumerable_type(enumerable_property_name: nil, current_json_object: nil)
+      if enumerable_property_name == :@job_triggers
+        type = current_json_object["type"]
+        # currently only supports 3 triggers
+        job_trigger = nil
+        if type == FastlaneCI::JobTrigger::TRIGGER_TYPE[:commit]
+          job_trigger = CommitJobTrigger.from_json!(current_json_object)
+        elsif type == FastlaneCI::JobTrigger::TRIGGER_TYPE[:nightly]
+          job_trigger = NightlyJobTrigger.from_json!(current_json_object)
+        elsif type == FastlaneCI::JobTrigger::TRIGGER_TYPE[:manual]
+          job_trigger = ManualJobTrigger.from_json!(current_json_object)
+        else
+          raise "Unable to parse JobTrigger type: #{type} from #{current_json_object}"
+        end
+        job_trigger
+      end
+    end
+
+    def self.json_to_attribute_name_proc_map
+      provider_object_to_provider = proc { |object|
+        provider_class = Object.const_get(object["class_name"])
+        if provider_class.include?(JSONConvertible)
+          provider = provider_class.from_json!(object)
+          provider
+        end
+      }
+      return { :@artifact_provider => provider_object_to_provider }
+    end
+
+    def self.attribute_name_to_json_proc_map
+      provider_to_provider_object = proc { |provider|
+        if provider.class.include?(JSONConvertible)
+          hash = provider.to_object_dictionary
+          hash
+        end
+      }
+      return { :@artifact_provider => provider_to_provider_object }
     end
   end
 end
